requirements.txt
Babel
click
filelock
python-fasthtml
python-dotenv
jsonschema
pydantic
pygeofilter
pygeoif
pyproj
python-dateutil
pytz
PyYAML
rasterio
httpx
shapely
SQLAlchemy
tinydb
